리팩토링: Custom hooks 'useCarousel'로 carousel 기능의 중복 코드 제거
