+++ src/App.js
@@ -2,7 +2,6 @@ import React from 'react';
 import Header from './components/header/Header';
 import Search from './components/search/Search';
 import Main from './components/main/Main';
-
 import BackGroundImg from './components/search/images/pexels-ready-made-3850607.jpg';
 import './App.css';
