+++ src/components/carousel/reviewCarousel/ReviewCarousel.js
@@ -1,38 +1,19 @@
-import React, { useState, useRef, useEffect } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import Profile from './Profile';
 import ItemInfo from '../ItemInfo';
 import CarouselBtn, { CarouselBtnWrap } from '../CarouselBtn';
 import ViewAll from '../VeiwAll';
+import useCarousel from '../../../hooks/useCarousel';
 
 const ReviewCarousel = ({ carouselInfo }) => {
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [leftPosition, setLeftPosition] = useState(0);
-  const [liClientWidth, setLiClientWidth] = useState(365);
-  const liSideMargin = 9;
-  const [liLength, setLiLength] = useState(5);
-  const refCarouselUl = useRef();
-  const refCarouselLi = useRef();
-
-  useEffect(() => {
-    setLiClientWidth(refCarouselLi.current?.clientWidth);
-    setLiLength(refCarouselUl.current?.childElementCount);
-  }, []);
-
-  const onCarouselBtnClick = (newIndex, newLeftPosition) => {
-    setCurrentIndex(newIndex);
-    setLeftPosition(newLeftPosition);
-  };
+  const { value, onCarouselBtnClick } = useCarousel();
+  const { refCarouselUl, refCarouselLi } = value;
 
   return (
     <Wrap>
       <h2>새로운 리뷰</h2>
-      <CarouselUl
-        ref={refCarouselUl}
-        leftPosition={leftPosition}
-        liClientWidth={liClientWidth}
-        liLength={liLength}
-        liSideMargin={liSideMargin}>
+      <CarouselUl ref={refCarouselUl} value={value}>
         {carouselInfo.map((info) => (
           <li key={info.id} ref={refCarouselLi}>
             <Profile userInfo={info.user} />
@@ -40,14 +21,7 @@ const ReviewCarousel = ({ carouselInfo }) => {
           </li>
         ))}
       </CarouselUl>
-      <CarouselBtn
-        currentIndex={currentIndex}
-        leftPosition={leftPosition}
-        liClientWidth={liClientWidth}
-        liSideMargin={liSideMargin}
-        liLength={liLength}
-        onCarouselBtnClick={onCarouselBtnClick}
-      />
+      <CarouselBtn value={value} onCarouselBtnClick={onCarouselBtnClick} />
       <ViewAll />
     </Wrap>
   );
@@ -72,14 +46,14 @@ const Wrap = styled.section`
 `;
 
 const CarouselUl = styled.ul`
-  width: ${(props) => (props.liClientWidth + props.liSideMargin * 2) * props.liLength}px;
+  width: ${(props) => (props.value.liClientWidth + props.value.liSideMargin * 2) * props.value.liLength}px;
   position: absolute;
-  left: ${(props) => props.leftPosition}px;
+  left: ${(props) => props.value.leftPosition}px;
   transition: all 0.3s ease;
 
   li {
     float: left;
     width: 366px;
-    margin: 0 ${(props) => props.liSideMargin}px 40px;
+    margin: 0 ${(props) => props.value.liSideMargin}px 40px;
   }
 `;
